Move `loadOrCreateDAO` into own file
